refactor(mobile): split accounts screen
- extract account list modules
- add screen source-contract test
