+++ apps/mobile/features/financial-accounts/components/financial-accounts-screen/FinancialAccountsScreen.types.ts
@@ -1,0 +1,7 @@
+import type { FinancialAccountRow } from "@/features/financial-accounts";
+
+export type FinancialAccountListItem = {
+  readonly account: FinancialAccountRow;
+  readonly identifiersCount: number;
+  readonly hasBillingProfileGap: boolean;
+};
